orq/app/requirements.txt
fastapi
uvicorn
pydantic
cassandra-driver
requests
